Basic worm god

+++ InSacrosanct/Assets/Scripts/LevelDesign/ContactDeath.cs
@@ -1,18 +1,17 @@
-using Cysharp.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.Timeline;
 
 public class ContactDeath : MonoBehaviour
-{    
+{
     [SerializeField]
     private TimelineAsset _deathTimeline;
-    
-    private bool _ended = false;
+
+    private bool _ended;
 
     private void OnTriggerEnter(Collider other)
     {
         var protag = other.GetComponentInParent<Protag>();
-        if (!_ended && protag)
+        if (!_ended && protag && enabled)
         {
             _ended = true;
             protag.Kill(_deathTimeline);
